refactor: Convert task add into a component
refactor: Convert task add into a component

* Moved all task add related code from Home.js to TaskAdd.js
* Imported TaskAdd into Home.js
* Home.js now uses <AddTask> component call
