+++ front-end/src/pages/TaskCard.css
@@ -8,7 +8,7 @@ body {
       monospace;
   }
 
-  /*-----TASK MAKE CARD-----*/
+  /*-----TASK ADD CARD-----*/
 
   .tasks-card {
    /*
